[SC-602] Solidity utils tech debt

+++ test/mocks/DaiLikePermitMock.sol
@@ -17,19 +17,23 @@ contract DaiLikePermitMock is ERC20Permit {
         _mint(initialAccount, initialBalance);
     }
 
-    function permit(address holder, address spender, uint256 nonce, uint256 expiry, bool allowed, uint8 v, bytes32 r, bytes32 s) external
-    {
-        bytes32 digest =
-            keccak256(abi.encodePacked(
+    function permit(
+        address holder,
+        address spender,
+        uint256 nonce,
+        uint256 expiry,
+        bool allowed,
+        uint8 v,
+        bytes32 r,
+        bytes32 s
+    ) external {
+        bytes32 digest = keccak256(
+            abi.encodePacked(
                 "\x19\x01",
                 this.DOMAIN_SEPARATOR(),
-                keccak256(abi.encode(PERMIT_TYPEHASH,
-                                    holder,
-                                    spender,
-                                    nonce,
-                                    expiry,
-                                    allowed))
-        ));
+                keccak256(abi.encode(PERMIT_TYPEHASH, holder, spender, nonce, expiry, allowed))
+            )
+        );
 
         require(holder != address(0), "Dai/invalid-address-0");
         require(holder == ecrecover(digest, v, r, s), "Dai/invalid-permit");
@@ -37,7 +41,7 @@ contract DaiLikePermitMock is ERC20Permit {
         require(expiry == 0 || block.timestamp <= expiry, "Dai/permit-expired");
         require(nonce == nonces(holder), "Dai/invalid-nonce");
         _useNonce(holder);
-        uint wad = allowed ? type(uint128).max : 0;
+        uint256 wad = allowed ? type(uint128).max : 0;
         _approve(holder, spender, wad);
         emit Approval(holder, spender, wad);
     }
